requirements.txt
aiogram==2.16
aiohttp==3.8.0
aiosignal==1.2.0
asyncpg==0.24.0
alembic==1.7.5
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
imdbpy==2021.4.18
importlib-resources==5.2.2
lxml==4.6.4
Mako==1.1.5
MarkupSafe==2.0.1
multidict==5.2.0
pytz==2021.3
requests==2.26.0
soupsieve==2.3.1
sqlalchemy==1.4.27
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.2
zipp==3.6.0
